refactor(runtime): extract message runtime
